back/requirements.txt
Flask
SQLAlchemy
Flask-SQLAlchemy
psycopg2-binary
requests
flask-cors
python-dotenv